python-master/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
google-auth==2.23.3
google-cloud-aiplatform==1.35.0
